requirements.txt
Django==1.11.23
django-restframework==0.0.1
pytz==2019.2
selenium==3.141.0
urllib3==1.25.3
